requirements.txt
azure-functions
pandas
jinja2
imgkit